+++ lib/types/index.d.ts
@@ -104,13 +104,14 @@ declare module 'react-native-wifi-reborn' {
      */
     export function connectToProtectedSSID(
         SSID: string,
-        password: string,
+        password: string | null,
         isWEP: boolean
     ): Promise<void>;
 
     export function getCurrentWifiSSID(): Promise<string>;
 
     //#region iOS only
+
     export function connectToSSID(SSID: string): Promise<void>;
     export function connectToSSIDPrefix(SSIDPrefix: string): Promise<void>;
     export function disconnectFromSSID(SSIDPrefix: string): Promise<void>;
@@ -119,47 +120,50 @@ declare module 'react-native-wifi-reborn' {
         password: string,
         isWEP: boolean
     ): Promise<void>;
+
     //#endregion
 
     //#region Android only
+
+    export interface WifiEntry {
+        SSID: string;
+        BSSID: number;
+        capabilities: string;
+        frequency: number;
+        level: number;
+        timestamp: number;
+    }
+
     /**
      * Returns a list of nearby WiFI networks.
      *
-     * @param callback Called if the attempt is successful. It contains a stringified JSONArray of `WiFiObject` as parameter.
-     * @param error Called if any error occurs during the attempt.
      * @example
-     * WifiManager.loadWifiList(
-            wifiList => {
-                let wifiArray =  JSON.parse(wifiList);
-                wifiArray.map((value, index) =>
-                    console.log(`Wifi ${index  +  1} - ${value.SSID}`)
-                );
-            },
-            error => console.log(error)
-        );
+     * const results = await WifiManager.loadWifiList();
+        results => {
+            let wifiArray =  JSON.parse(results);
+            wifiArray.map((value, index) =>
+                console.log(`Wifi ${index  +  1} - ${value.SSID}`)
+            );
+        },
      */
-    export function loadWifiList(
-        callback: (wifiList: string) => void,
-        error: (err: string) => void
-    ): void;
+    export function loadWifiList(): Promise<Array<WifiEntry>>;
 
     /**
-     * Similar to `loadWifiList` but it forcefully starts the WiFi scanning on android and in the callback fetches the list.
+     * Similar to `loadWifiList` but it forcefully starts a new WiFi scan and only passes the results when the scan is done.
      */
-    export function reScanAndLoadWifiList(
-        callback: (wifiList: string) => void,
-        error: (err: string) => void
-    ): void;
+    export function reScanAndLoadWifiList(): Promise<Array<string>>;
 
-    export function isEnabled(callback: (enabled: boolean) => void): void;
+    /**
+     * Method to check if wifi is enabled.
+     */
+    export function isEnabled(): Promise<boolean>;
 
     export function setEnabled(enabled: boolean): void;
 
     /**
-     * Indicates whether network connectivity exists and it is possible to establish connections.
-     * @param Called when the network status is resolved.
+     * Returns if the device is currently connected to a WiFi network.
      */
-    export function connectionStatus(callback: (isConnected: boolean) => void): void;
+    export function connectionStatus(): Promise<boolean>;
 
     export function disconnect(): void;
 
@@ -167,6 +171,26 @@ declare module 'react-native-wifi-reborn' {
         couldNotGetWifiManager = 'couldNotGetWifiManager',
         couldNotGetConnectivityManager = 'couldNotGetConnectivityManager',
     }
+
+    /**
+     * Returns the BSSID (basic service set identifier) of the currently connected WiFi network.
+     */
+    export function getBSSID(): Promise<string>;
+
+    /**
+     * Returns the RSSI (received signal strength indicator) of the currently connected WiFi network.
+     */
+    export function getCurrentSignalStrength(): Promise<number>;
+
+    /**
+     * Returns the frequency of the currently connected WiFi network.
+     */
+    export function getFrequency(): Promise<number>;
+
+    /**
+     * Returns the IP of the currently connected WiFi network.
+     */
+    export function getIP(): Promise<string>;
 
     /**
      * This method will remove the wifi network configuration.
@@ -192,5 +216,6 @@ declare module 'react-native-wifi-reborn' {
      * @param useWifi boolean to force wifi off or on
      */
     export function forceWifiUsage(useWifi: boolean): Promise<void>;
+
     //#endregion
 }